two minor code changes plus German documentation

+++ src/Look/InfoBoxLook.cpp
@@ -26,7 +26,6 @@ Copyright_License {
 #include "Colors.hpp"
 #include "Screen/Layout.hpp"
 #include "AutoFont.hpp"
-#include "Interface.hpp"
 #include "Util/Clamp.hpp"
 
 #ifdef HAVE_TEXT_CACHE
@@ -78,7 +77,7 @@ void
 InfoBoxLook::ReinitialiseLayout(unsigned width, unsigned scale_title_font)
 {
   FontDescription title_font_d(8);
-  unsigned scaled_width = (width * Clamp((int)scale_title_font,50,150)) / 100;
+  unsigned scaled_width = (width * Clamp(scale_title_font,50U,150U)) / 100;
   AutoSizeFont(title_font_d, scaled_width, _T("123456789012345"));
   title_font.Load(title_font_d);
